+++ src/main/java/core/information/ParentInformation.java
@@ -1,4 +1,4 @@
-package core.information2;
+package core.information;
 
 import org.jetbrains.annotations.Contract;
 import org.jetbrains.annotations.NotNull;
