+++ src/test/java/core/information/DatamodelNodesTest.java
@@ -1,4 +1,4 @@
-package core.information2;
+package core.information;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -7,7 +7,7 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static util.DataGenerationUtil2.*;
+import static util.DataGenerationUtil.*;
 
 public class DatamodelNodesTest {
     Ref<ProjectInformation, RootInformation> proj;
