+++ src/test/java/core/information/DatamodelRelationsTest.java
@@ -1,10 +1,10 @@
-package core.information2;
+package core.information;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static util.DataGenerationUtil2.*;
+import static util.DataGenerationUtil.*;
 
 @SuppressWarnings({"unused", "SpellCheckingInspection"})
 public class DatamodelRelationsTest {
